fs references are now removed from js api file
