+++ src/api/0.1/cli/aws.js
@@ -2,12 +2,14 @@
 
 const cn = require('../js/js-api');
 const cmn = require('../common');
+const fs = require('../project-fs/fs');
 
 const clusternatorJson = cmn.src('clusternator-json');
 
 module.exports = {
   listProjects: cn.listProjects,
-  describeServices
+  describeServices,
+  certUpload
 };
 
 /**
@@ -19,4 +21,18 @@ function describeServices() {
     .get()
     .then((config) => cn
       .describeProject(config.projectId));
+}
+
+/**
+ * @param {string} privateKey
+ * @param {string} certificate
+ * @param {string} certId
+ * @param {string=} chain
+ * @return {Q.Promise}
+ */
+function certUpload(privateKey, certificate, certId, chain) {
+  return fs
+    .loadCertificateFiles(privateKey, certificate, chain)
+    .then((certs) => cn.certUpload(certId, certs));
+
 }
